DBA-007: Update init-hooks.sh to check if files exists first

+++ .gitlab/_scripts/init-hooks.sh
@@ -1,4 +1,8 @@
 #!/usr/bin/env sh
 
-ln -s ../../.gitlab/_hooks/commit-msg .git/hooks
-ln -s ../../.gitlab/_hooks/pre-push .git/hooks
+if [ ! -f .git/hooks/commit-msg ]; then
+    ln -s ../../.gitlab/_hooks/commit-msg .git/hooks
+fi
+if [ ! -f .git/hooks/pre-push ]; then
+    ln -s ../../.gitlab/_hooks/pre-push .git/hooks
+fi
